+++ src/command_modules/azure-cli-advisor/azure/cli/command_modules/advisor/tests/latest/test_advisor_commands.py
@@ -5,11 +5,13 @@
 
 import unittest
 
-from azure.cli.testsdk import LiveScenarioTest, ScenarioTest, ResourceGroupPreparer
+from azure.cli.testsdk import (
+    LiveScenarioTest, ScenarioTest, ResourceGroupPreparer)
 
-from azure.cli.command_modules.advisor.custom import build_filter_string, parse_operation_id, parse_recommendation_uri
-
-# pylint: disable=line-too-long
+from azure.cli.command_modules.advisor.custom import (
+    _build_filter_string,
+    _parse_operation_id,
+    _parse_recommendation_uri)
 
 
 class AzureAdvisorUnitTest(unittest.TestCase):
@@ -19,81 +21,173 @@ class AzureAdvisorUnitTest(unittest.TestCase):
         resource_group_name = 'r'
         category = 'cost'
 
-        self.assertEqual(build_filter_string(), None)
-        self.assertEqual(build_filter_string(ids=ids), "ResourceId eq 'a' or ResourceId eq 'b' or ResourceId eq 'c'")
-        self.assertEqual(build_filter_string(resource_group_name=resource_group_name), "ResourceGroup eq 'r'")
-        self.assertEqual(build_filter_string(category=category), "Category eq 'cost'")
-        self.assertEqual(build_filter_string(ids=ids, category=category), "(ResourceId eq 'a' or ResourceId eq 'b' or ResourceId eq 'c') and Category eq 'cost'")
-        self.assertEqual(build_filter_string(resource_group_name=resource_group_name, category=category), "(ResourceGroup eq 'r') and Category eq 'cost'")
+        self.assertEqual(
+            _build_filter_string(),
+            None)
+        self.assertEqual(
+            _build_filter_string(ids=ids),
+            "ResourceId eq 'a' or ResourceId eq 'b' or ResourceId eq 'c'")
+        self.assertEqual(
+            _build_filter_string(
+                resource_group_name=resource_group_name),
+            "ResourceGroup eq 'r'")
+        self.assertEqual(
+            _build_filter_string(
+                category=category),
+            "Category eq 'cost'")
+        self.assertEqual(
+            _build_filter_string(
+                ids=ids,
+                category=category),
+            "(ResourceId eq 'a' or ResourceId eq 'b' or ResourceId eq 'c') and Category eq 'cost'")
+        self.assertEqual(
+            _build_filter_string(
+                resource_group_name=resource_group_name,
+                category=category),
+            "(ResourceGroup eq 'r') and Category eq 'cost'")
 
-    def test_parse_operation_id(self):
-        location = 'https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Advisor/generateRecommendations/a9544ca5-5837-4cb4-94d6-bad2b6e76320?api-version=2017-04-19'
-        operation_id = parse_operation_id(location)
+    def test__parse_operation_id(self):
+        location = ("https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/"
+                    "providers/Microsoft.Advisor/generateRecommendations/a9544ca5-5837-4cb4-94d6-bad2b6e76320"
+                    "?api-version=2017-04-19")
+        operation_id = _parse_operation_id(location)
         self.assertEqual(operation_id, 'a9544ca5-5837-4cb4-94d6-bad2b6e76320')
 
-    def test_parse_recommendation_uri(self):
-        recommendation_uri = '/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Network/expressRouteCircuits/test/providers/Microsoft.Advisor/recommendations/c4deb869-ea38-f90d-331f-91770021d425'
-        result = parse_recommendation_uri(recommendation_uri)
+    def test__parse_recommendation_uri(self):
+        recommendation_uri = ("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo"
+                              "/providers/Microsoft.Network/expressRouteCircuits/test/"
+                              "providers/Microsoft.Advisor/recommendations/c4deb869-ea38-f90d-331f-91770021d425"
+                              "/suppressions/5c9c3fce-c1b2-7e45-106c-152ce3c04be5")
+        result = _parse_recommendation_uri(recommendation_uri)
         self.assertEqual(
-            result['resourceUri'],
-            '/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Network/expressRouteCircuits/test'
-        )
-        self.assertEqual(result['recommendationId'], 'c4deb869-ea38-f90d-331f-91770021d425')
+            result['resource_uri'],
+            ("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo"
+             "/providers/Microsoft.Network/expressRouteCircuits/test"))
+        self.assertEqual(result['recommendation_id'], 'c4deb869-ea38-f90d-331f-91770021d425')
 
 
 class AzureAdvisorLiveScenarioTest(LiveScenarioTest):
 
-    def test_list_disable_enable_recommendations(self):
-        self.cmd('advisor recommendation generate')
-        output = self.cmd('advisor recommendation list').get_output_in_json()
-        self.assertGreater(len(output), 1)
-        output = self.cmd('advisor recommendation list --category cost').get_output_in_json()
-        self.assertGreater(len(output), 0)
+    def test_recommendations(self):
+        # List should return at least one recommendation with filters
+        output = self.cmd('advisor recommendation list --category Security',
+                          checks=self.check("[0].category", 'Security')).get_output_in_json()
 
+        # Set the recommendation details to use with further commands
+        rec = output[0]
+        name = rec['name']
+        recommendation_id = rec['id']
+        resource_id = _parse_recommendation_uri(recommendation_id)
         self.kwargs.update({
-            'recommendation_id': output[0]['id']
+            'name': name,
+            'recommendation_id': recommendation_id,
+            'resource_id': resource_id
         })
-        disableCmd = 'advisor recommendation disable --ids {recommendation_id} --days 1'
-        enableCmd = 'advisor recommendation enable --ids {recommendation_id}'
-        output = self.cmd(disableCmd).get_output_in_json()
-        self.assertEqual(output[0]['ttl'], '1.00:00:00')
-        self.cmd(enableCmd)
-        output = self.cmd('advisor recommendation list --category cost').get_output_in_json()
-        self.assertGreater(len(output), 0)
-        self.assertEqual(output[0]['suppressionIds'], None)
+
+        # Disable with specified duration should create a suppression
+        self.cmd('advisor recommendation disable --ids {recommendation_id} --days 1',
+                 checks=[self.check("[0].id", recommendation_id),
+                         self.exists("[0].suppressionIds")])
+
+        # Disable again should create another suppression
+        self.cmd('advisor recommendation disable --ids {recommendation_id}',
+                 checks=[self.check("[0].id", recommendation_id),
+                         self.greater_than("length([0].suppressionIds)", 1)])
+
+        # Enable should remove all suppressions
+        self.cmd('advisor recommendation enable --ids {recommendation_id}',
+                 checks=[self.check("[0].id", recommendation_id),
+                         self.check("[0].suppressionIds", None)])
+
+        # Enable again should be a no op
+        self.cmd('advisor recommendation enable --ids {recommendation_id}',
+                 checks=[self.check("[0].id", recommendation_id),
+                         self.check("[0].suppressionIds", None)])
+
+    def test_recommendations_resource_group(self):
+        resource_group = 'AzExpertStg'
+        self.kwargs.update({
+            'resource_group': resource_group
+        })
+
+        # List should return at least one recommendation with filters
+        output = self.cmd('advisor recommendation list --resource-group {resource_group}',
+                          checks=self.check("[0].resourceGroup", resource_group)).get_output_in_json()
+
+        # Set the recommendation details to use with further commands
+        resource_id = _parse_recommendation_uri(output[0]['id'])['resource_uri']
+        name = output[0]['name']
+        self.kwargs.update({
+            'resource_id': resource_id,
+            'name': name
+        })
+
+        # Disable with specified duration should create a suppression
+        self.cmd('advisor recommendation disable -n {name} -g {resource_group} --days 1',
+                 checks=[self.check("[0].name", name),
+                         self.check("[0].resourceGroup", resource_group),
+                         self.exists("[0].suppressionIds")])
+
+        # Disable again should create another suppression
+        self.cmd('advisor recommendation disable -n {name} -g {resource_group}',
+                 checks=[self.check("[0].name", name),
+                         self.check("[0].resourceGroup", resource_group),
+                         self.greater_than("length([0].suppressionIds)", 1)])
+
+        # Enable should remove all suppressions
+        self.cmd('advisor recommendation enable -n {name} -g {resource_group}',
+                 checks=[self.check("[0].name", name),
+                         self.check("[0].resourceGroup", resource_group),
+                         self.check("[0].suppressionIds", None)])
+
+        # Enable again should be a no op
+        self.cmd('advisor recommendation enable -n {name} -g {resource_group}',
+                 checks=[self.check("[0].name", name),
+                         self.check("[0].resourceGroup", resource_group),
+                         self.check("[0].suppressionIds", None)])
 
 
 class AzureAdvisorScenarioTest(ScenarioTest):
 
-    def test_generate_recommendations(self):
-        self.cmd('advisor recommendation generate')
+    def test_configurations_subscription(self):
+        # Show should always return a default
+        self.cmd('advisor configuration show',
+                 checks=[self.check("type", "Microsoft.Advisor/Configurations")])
 
-    def test_get_set_configurations_subscription(self):
-        output = self.cmd('advisor configuration get').get_output_in_json()
-        self.assertGreater(len(output), 1)
-        self.cmd('advisor configuration set --low-cpu-threshold 20')
-        output = self.cmd('advisor configuration get').get_output_in_json()
-        for entry in output['value']:
-            if entry['properties']['lowCpuThreshold']:
-                self.assertEqual(entry['properties']['lowCpuThreshold'], '20')
-        self.cmd('advisor configuration set --low-cpu-threshold 5')
-        output = self.cmd('advisor configuration get').get_output_in_json()
-        for entry in output['value']:
-            if entry['properties']['lowCpuThreshold']:
-                self.assertEqual(entry['properties']['lowCpuThreshold'], '5')
+        # Show should reflect the changes made by Update
+        self.cmd('advisor configuration update --low-cpu-threshold 15 --exclude')
+        self.cmd('advisor configuration show',
+                 checks=[self.check("properties.lowCpuThreshold", "15"),
+                         self.check("properties.exclude", True)])
+
+        # Show should reflect the changes made by Update
+        self.cmd('advisor configuration update --low-cpu-threshold 5 --include')
+        self.cmd('advisor configuration show',
+                 checks=[self.check("properties.lowCpuThreshold", "5"),
+                         self.check("properties.exclude", False)])
+
+        # List should reflect the changes made by Update
+        self.cmd('advisor configuration list',
+                 checks=[self.check("[0].properties.lowCpuThreshold", "5"),
+                         self.check("[0].properties.exclude", False)])
 
     @ResourceGroupPreparer(name_prefix='cli_test_advisor')
-    def test_get_set_configurations_resource_group(self, resource_group):
-        output = self.cmd('advisor configuration get --resource-group {rg}').get_output_in_json()
-        self.assertGreater(len(output), 0)
-        self.cmd('advisor configuration set --exclude --resource-group {rg}')
-        output = self.cmd('advisor configuration get --resource-group {rg}').get_output_in_json()
-        self.assertGreater(len(output), 0)
-        self.assertTrue(output['value'][0]['properties']['exclude'])
-        self.cmd('advisor configuration set --include --resource-group {rg}')
-        output = self.cmd('advisor configuration get --resource-group {rg}').get_output_in_json()
-        self.assertGreater(len(output), 0)
-        self.assertTrue(not output['value'][0]['properties']['exclude'])
+    def test_configurations_resource_group(self, resource_group):
+        # Show should always return a default even for a brand new resource group
+        self.cmd('advisor configuration show -g {rg}',
+                 checks=[self.check("resourceGroup", resource_group)])
+
+        # Show should reflect the changes made by Update
+        self.cmd('advisor configuration update --exclude -g {rg}')
+        self.cmd('advisor configuration show -g {rg}',
+                 checks=[self.check("resourceGroup", resource_group),
+                         self.check("properties.exclude", True)])
+
+        # Show should reflect the changes made by Update
+        self.cmd('advisor configuration update --include -g {rg}')
+        self.cmd('advisor configuration show -g {rg}',
+                 checks=[self.check("resourceGroup", resource_group),
+                         self.check("properties.exclude", False)])
 
 
 if __name__ == '__main__':