+++ src/command_modules/azure-cli-advisor/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.5.0
+++++++
+* BC: `advisor configuration get` has been renamed to `advisor configuration list`.
+* BC: `advisor configuration set` has been renamed to `advisor configuration update`.
+* BC: `advisor recommendation generate` has been removed.
+* `advisor recommendation list` has a new --refresh parameter.
+* `advisor recommendation show` has been added.
+
 0.1.2
 +++++
 * Minor fixes.